deleted useless meta files
Deleted various files that aren't really needed for other machines,
some files (pyproject) are purely used for formatting and testing and
other files (.devenv) signify that i am updating on a development
machine which for generic users won't be the case.

Also TODO: add a way for quotes.json to be generated automatically.